pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentify-toolkit"
version = "0.1.0"
description = "Declarative AI toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Lewis Sheridan", email="lewis@backplane.dev" }
]

keywords = ["ai", "agents", "llm", "openai", "anthropic", "bedrock"]

classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
]

dependencies = [
  "pyyaml>=6.0",
  "rich>=13.0",
  "click>=8.1"
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
bedrock = ["boto3>=1.34"]
grok = ["xai_sdk>=1.5.0"]

[project.scripts]
agentify = "agentify.cli:main"

[project.urls]
Homepage = "https://github.com/backplane-software/agentify"
Issues = "https://github.com/backplane-software/agentify/issues"